requirements/base.txt
APScheduler==3.6.3
asgiref==3.5.0
backports.zoneinfo==0.2.1
blessings==1.7
bpython==0.22.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
curtsies==0.3.10
cwcwidth==0.1.6
Django==4.0.3
greenlet==1.1.2
idna==3.3
psycopg2-binary==2.9.3
Pygments==2.11.2
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.27
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tornado==6.1
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.9
